Add suffix argument
Extend the funcitonality of the function: A suffix can be specified
which is add to either the original filename or in case the rename
argument is not None to the date format string. If the suffix is
either the string "File" or "file", the original filename is used
as suffix.