add prefix feature and pardir option for file renaming
Prefix feature: option to add prefix to filename. Could be orginal filename
in case File/file is provided to the option or the original parent directories name in case
Pardir/pardir is provided. Also any other string can be used as prefix.
The parentdirectory option is also available for suffixes now.